requirements.txt
streamlit==1.38.0
langchain==0.2.16
langgraph==0.2.14
langchain-groq==0.1.9
finnhub-python==2.4.19
newsapi-python==0.2.7
mysql-connector-python==8.3.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.9.2
python-dotenv==1.0.1
bcrypt==4.2.0

streamlit==1.38.0
langchain==0.3.0
langchain-groq==0.2.0
langgraph==0.2.14
pydantic==2.9.2
mysql-connector-python==8.3.0
finnhub-python==2.4.19
bcrypt==4.2.0
pandas==2.2.2